coding/langchain/deepagent-langchain/Chapter_2_Quickstart/requirements.txt
deepagents
tavily-python
langchain-openai
python-dotenv
rich
